// dotnet/openapi/model/jobModelServiceResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobModel } from './jobModel';


export interface JobModelServiceResponse { 
    message?: string | null;
    statusCode?: number;
    data?: JobModel;
}