// dotnet/openapi/model/updateJobCommand.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateJobCommand { 
    title?: string | null;
    description?: string | null;
    location?: string | null;
    skillsRequired?: string | null;
    experienceRequired?: number;
    jobType?: string | null;
    salary?: number;
    expiryDate?: string;
    education?: string | null;
    clientCompanyId?: string;
    id?: string;
}